// big-fish-lib/types/checkouts.ts
export type SingleFieldId =
  | '1'
  | '2'
  | '3'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'
  | '9'
  | '10'
  | '11'
  | '12'
  | '13'
  | '14'
  | '15'
  | '16'
  | '17'
  | '18'
  | '19'
  | '20'
  | 'B'

export type DoubleFieldId =
  | 'D1'
  | 'D2'
  | 'D3'
  | 'D4'
  | 'D5'
  | 'D6'
  | 'D7'
  | 'D8'
  | 'D9'
  | 'D10'
  | 'D11'
  | 'D12'
  | 'D13'
  | 'D14'
  | 'D15'
  | 'D16'
  | 'D17'
  | 'D18'
  | 'D19'
  | 'D20'
  | 'DB'

export type TripleFieldId =
  | 'T1'
  | 'T2'
  | 'T3'
  | 'T4'
  | 'T5'
  | 'T6'
  | 'T7'
  | 'T8'
  | 'T9'
  | 'T10'
  | 'T11'
  | 'T12'
  | 'T13'
  | 'T14'
  | 'T15'
  | 'T16'
  | 'T17'
  | 'T18'
  | 'T19'
  | 'T20'

export type AnyFieldId = SingleFieldId | DoubleFieldId | TripleFieldId

export type CheckoutInOne = [DoubleFieldId]
export type CheckoutInTwo = [AnyFieldId, DoubleFieldId]
export type CheckoutInThree = [AnyFieldId, AnyFieldId, DoubleFieldId]

export type Checkout = { id: string; value: number; sequence: CheckoutInOne | CheckoutInTwo | CheckoutInThree }